requirements.txt
yfinance>=0.2.50
pandas>=2.2.0
numpy>=2.0.0
scikit-learn>=1.7.2
xgboost>=2.1.0
torch>=2.5.0
mlflow>=2.18.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
boto3>=1.35.0
evidently>=0.4.30
pyyaml>=6.0.2
python-dotenv>=1.0.1
httpx>=0.27.0
joblib>=1.4.0
ta>=0.11.0
